requirements.txt
langchain>=0.3.27
langchain-core>=0.1.31
langchain-community>=0.0.24
langgraph>=0.6.7
langchain-openai>=0.3.33
playwright>=1.55.0
qdrant-client>=1.15.1
fastapi>=0.117.1
uvicorn>=0.37.0
python-dotenv>=1.1.1
atlassian-python-api>=4.0.7
PyGithub>=2.8.1
ragas>=0.3.5
lxml>=6.0.2
beautifulsoup4>=4.13.5
sentence-transformers>=5.1.1
pytest>=8.4.2
black>=25.9.0
cryptography>=42.0.5
difflib3>=0.1.1
python-jose>=3.3.0
jq>=1.6.0
plotly>=5.19.0
jinja2>=3.1.3
markdown2>=2.4.13